{"schema":"decodex/commit/1","summary":"Modularize worktree git and hook helpers","authority":"manual"}
